+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+### **(V.4.0.1 Changes) (1.18.1 Minecraft)**
+
+##### Blocks:
+Fixed Honey Fluid bottom not rendering when it is slowly falling onto an opaque block.
+
+Fixed Honey Fluid falling faster if there is neighboring Honey Fluid that is also falling.
+
+##### Mod Compat:
+Fixed crash with some mods asking for getBaseHeight from Bumblezone's chunk generator.
+
+
 ### **(V.4.0.0 Changes) (1.18.1 Minecraft)**
 
 ##### Major: